Refactor createPanel into createPanel.js 37 / 66 tests work.
