+++ src/hud/core/hudCommon.js
@@ -19,27 +19,7 @@ export function ensureHudRoot() {
   return panels;
 }
 
-/**
- * @param {{id?: string, title: string}} opts
- * @returns {{root:HTMLElement, header:HTMLElement, body:HTMLElement}}
- */
-export function createPanel({ id, title }) {
-  const root = document.createElement("div");
-  if (id) root.id = id;
-  root.className = "hud-panel";
 
-  const header = document.createElement("div");
-  header.className = "hud-title";
-  header.textContent = title;
-
-  const body = document.createElement("div");
-  body.className = "hud-body";
-
-  root.appendChild(header);
-  root.appendChild(body);
-
-  return { root, header, body };
-}
 
 export function mountPanel(panel, container = ensureHudRoot()) {
   container.appendChild(panel.root);
