+++ src/hud/panels/cityGeneratorHud.js
@@ -1,20 +1,11 @@
-import { ensureHudRoot, createPanel, mountPanel, destroyPanel } from "../core/hudCommon.js";
+import { ensureHudRoot, mountPanel, destroyPanel } from "../core/hudCommon.js";
 import {createButton} from "../elements/createButton.js";
+import {makeSlider} from "../elements/createSlider.js";
+import {createPanel} from "../elements/createPanel.js";
 
 export function createCityGeneratorHUD({ getWorld, getRegistry } = {}) {
   const panel = createPanel({ id: "hud-city-gen", title: "City Generator" });
 
-  function makeSlider(label, min, max, step, get, set) {
-    const row = document.createElement("div");
-    row.style.margin = "6px 0";
-    const l = document.createElement("div"); l.textContent = label; l.style.fontSize="12px"; l.style.opacity="0.85";
-    const s = document.createElement("input"); s.type = "range"; s.min = String(min); s.max = String(max); s.step = String(step);
-    s.value = String(get());
-    const v = document.createElement("span"); v.style.marginLeft = "8px"; v.textContent = s.value;
-    s.oninput = () => { v.textContent = s.value; set(parseFloat(s.value)); };
-    row.appendChild(l); row.appendChild(s); row.appendChild(v);
-    return row;
-  }
 
   function regen() {
     const w = getWorld?.(); const r = getRegistry?.();
