+++ src/hud/panels/keyOverlayHud.js
@@ -1,4 +1,5 @@
-import { createPanel, mountPanel, destroyPanel, ensureHudRoot } from "../core/hudCommon.js";
+import {  mountPanel, destroyPanel, ensureHudRoot } from "../core/hudCommon.js";
+import {createPanel} from "../elements/createPanel.js";
 
 const KEYS = [
   { code: "KeyQ", label: "Q" },
